Make suppression tags configurable (#14)
In our use-case, we want to suppress running mypy using either the tag `no-mypy` or the tag `no-lint`. Before we open-sourced this ruleset, we used an "eligibility function", but almost always configured that using rule kind and tags, so we're opting here for the narrower (and simpler to perform) customization.