Add rule's data to mypy action
This allows a py_library target with a native extension to propagate its
`pyi` file so that it is considered by mypy.

This has the potential downside of including more files in the action
than necessary, since not all data might be used by mypy. If that's a
problem we could likely filter this by filename