Hacking on the $MYPYPATH behavior
It appears that `rules_mypy` has several issues interacting with MyPy,
specifically regarding how Bazel wants to lay out runfiles.

MyPy has a multi-tier path search mechanism, see `mypy/modulefinder.py`
for details. The TLDR is that it has integrated knowledge of how search
paths should be set up defaulting to

    search_paths = SearchPaths(
        python_path=(os.getcwd(),),
        mypy_path=tuple(mypy_path() + options.mypy_path),
        package_path=tuple(sys_path),
        typeshed_path=(),
    )

When MyPy searches for a module, it prefers entries in the MYPYPATH to
entries either in the local source tree or in the virtualenv/PYTHONPATH.

Notionally this should be all well and good. `rules_python` already
handles setting up a `PYTHONPATH`, all we should have to do is copy
import roots that would be `PYTHONPATH` entries under `rules_python`
into the `MYPYPATH`. Ordinarily `rules_python` will do this for us, but
since `rules_mypy` runs MyPy as a tool outside of the Python environment
for the target being checked we have to implement that ourselves. And
this works fine for both simple 3rdparty code and 1stparty code.

The trick is how Bazel treats generated files, which don't land in the
logical source tree but instead land in the `bazel-bin/` tree even
within execroots. Consider the following example

    genrule(
        name = "generate_source",
        outs = ["bar.py"],
        cmd = "echo 'def foo(): print(\"hi\")' > $(OUTS)",
    )

    py_library(
        name = "bar",
        srcs = ["bar.py"],
        imports = ["."],
        visibility = ["//visibility:public"],
    )

Here the `py_library` will have a PyInfo provider which lists the
package as an import root which is all well and good, but the `bar.py`
generated file won't land in the package. It'll land in a parallel
`bazel-bin` tree which isn't a child of the import anymore.

The existence of this parallel tree creates other problems too because
even if the source and generated roots are listed as `MYPYPATH` entries
there's potential for difficult to resolve path ordering problems in the
presence of `__init__.py` files, generated or otherwise.

While this patch is better than the stock `rules_mypy` behavior in that
it eliminates the previous cross product of import roots and import
suffixes, this still isn't quite a correct implementation of the
required logic.

A fully correct implementation likely needs to traverse all the files
required for a given typecheck, identify import paths, generate a
`MYPYPATH` consisting only of source-tree native import paths, and then
traverse all the files creating `--shadow-file` mappings to redirect
MyPy to the generated file's location. This would avoid having
effectively duplicate path entries and the problems that can come along
with it.

Work towards theorempl/rules_mypy#74