Avoid excessive BES event size
The files in requested output groups are announced as events on the Build Event Stream (BES), with declared directories expanded into the list of files they contain. Since mypy caches contain information about all transitive dependencies, including the cache directory in the `mypy` output group thus resulted in a very large number of (duplicated) files announced in build events when the aspect is applied to an entire repo, causing excessive upload times and BES sizes in the GBs.

This is fixed by dropping the cache directory from the `mypy` output group. It is still accessible via the `MypyCacheInfo` provider.