Fix mypy linting non-python source files with custom rules
If you write a custom rule that is designed to interop with python rules
you need to return PyInfo, in this case rules_mypy was attempting to
lint the source files of the target even if they were not python. This
is now scoped to only `.py` and `.pyi` files.

One example case for this is if you write a custom rule that creates a
python native extension and you want to be able to add it to the `deps`
of python rules.