Update mypy config file examples to use Label
Some mypy label references are used between repos (main and rules_mypy) so it is important that they are declared relative to the original bzl file correctly.  Use `Label()` at the construction site to do this.

(This is the "correct" fix for needing the full canonical name in string form, so update the advice to use "@@//name")